Update : refactoring. move fixture to conftest
